backup/pyproject.toml
[project]
name = "chilean-humor"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Alonso Astroza Tagle", email = "alonsoastroza@gmail.com" }
]
dependencies = [
    "datasette>=0.64.6",
    "pandas>=2.2.2",
    "jupyter>=1.0.0",
    "seaborn>=0.13.2",
    "sqlite-utils>=3.36",
    "youtube-transcript-api>=0.6.2",
    "loguru>=0.7.2",
    "openai>=1.30.3",
    "instructor>=1.3.1",
    "pgvector>=0.2.5",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "pytube>=15.0.0",
    "wordcloud>=1.9.3",
    "nltk>=3.8.1",
    "spacy>=3.7.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/chilean_humor"]
